Revert to sh-sdhi driver
In commit 9ac9c341f9 the sh-sdhi driver was replaced with the common
Uniphier driver (for Matsushita SD IP) on R-Car Gen3 SoCs. With the
common driver, reading from eMMC is unstable and results in frequent boot
failures on some hardware, probably due to a timing/race condition issue.

As far as can be observed, the old sh-sdhi driver does not trigger
this issue. Revert to the old driver for now for reliably boot until the
underlying issue can be diagnosed.