usb: hub: Continue scan on port when disconnect is reported
For some device/hub combinations a "spurious" disconnect event was
observed when the hub turned the port on. The logic in usb_scan_port
since f7a9e5dd interpreted this as a new USB device which it tried (and
failed) to activate, and then removed the port from the scan list
resulting in the device attached to the port not being usable.

Do not attempt to add a device and do not remove the port from the scan
list when a connection status change but no connection is reported,
instead wait for the device to reappear.